refactor: move tests for project member validation to ProjectMemberValidatorTest
